+++ resources/lang/de/changelog.php
@@ -1,6 +1,0 @@
-<?php
-
-return [
-    'changelog' => 'Änderungshistorie',
-    'changelog_link_desc' => 'Lies mehr über die neusten Änderungen in dieser Applikation <a href=":link">hier</a>.',
-];