+++ resources/lang/en/changelog.php
@@ -1,6 +1,0 @@
-<?php
-
-return [
-    'changelog' => 'Changelog',
-    'changelog_link_desc' => 'Read about the latest changes of this application <a href=":link">here</a>.',
-];