+++ resources/views/changelog.blade.php
@@ -1,0 +1,7 @@
+@extends('layouts.app')
+
+@section('title', __('app.changelog'))
+
+@section('content')
+    {!! $content !!}
+@endsection