+++ resources/views/changelog/index.blade.php
@@ -1,9 +1,0 @@
-@extends('layouts.app')
-
-@section('title', __('changelog.changelog'))
-
-@section('content')
-
-    {!! $changelog !!}
-
-@endsection